Euclidean and Binary methods implemented
