+++ gcd/gcd.cpp
@@ -1,40 +1,94 @@
-/*
- * 
- * Program to calculate the greatest common divisor of 
- * two given numbers using two methods - 
- * 1) Euclidean method of long division
- * 2) Binary algorithm, uses mainly divsion and subtraction by 2
- * 
-*/
+// Implementation of the functions needed for gcd calculation
 
-#include <iostream>
-#include <string>
+// #include <iostream>
 #include "gcd.h"
 
-int main()
+int max(int a, int b)
 {
-	int a, b;
-	int g = 0;
+	int m = (a > b) ? a : b;
+	return m;
+}
+
+int min(int a, int b)
+{
+	int m = (a < b) ? a : b;
+	return m;
+}
+
+
+int gcd_binary(int a, int b)
+{
+	int c, d = 0;
 	
-	std::string outstr{""};
+	if(a == b)
+		return a;
 	
-	// The two integers
-	std::cout << "Enter two integers" << '\n';
-	std::cin >> a >> b;
+	if(a == 0)
+		return b;
 	
-	// 1. Euclidean method
-	std::cout << "Euclidean method: " << '\n';
-	g = gcd_euclid(a, b);
+	if(b == 0)
+		return a;
 	
-	// put the result out
-	std::cout << "GCD(" << a << " , " << b << ") = " << g << '\n';
+	if(a%2 == 0 && b%2 == 0)
+	{
+		// divide both by 2 and recurse
+		a /= 2;
+		b /= 2;
+		//std::cout << "(" << a << "," << b << ")\n";
+		return 2*gcd_binary(a, b);
+	}
+	else
+	{
+		// at least one of them is odd
+		if(a%2 == 0 && b%2 != 0)
+		{
+			a /= 2;
+			//std::cout << "(" << a << "," << b << ")\n";
+			return gcd_binary(a, b);
+		}
+		if(a%2 !=0 && b%2 == 0)
+		{
+			b /= 2;
+			//std::cout << "(" << a << "," << b << ")\n";
+			return gcd_binary(a, b);
+		}
+		if(a%2 !=0 && b%2 != 0)
+		{
+			if(a > b)
+			{
+				c = (a - b)/2;
+				//std::cout << "(" << a << "," << b << ")\n";
+				return gcd_binary(c, b);
+			}
+			else
+			{
+				c = (b - a)/2;
+				//std::cout << "(" << a << "," << b << ")\n";
+				return gcd_binary(c, a);
+			}
+		}
+	}
+}
+
+
+int gcd_euclid(int a, int b)
+{
+	int r, rm;
+	int n1, n2;
 	
-	// 2. Binary method
-	std::cout << "Binary method: " << '\n';
-	g = gcd_binary(a, b);
+	n1 = min(a, b);
+	n2 = max(a, b);
 	
-	// put the result out
-	std::cout << "GCD(" << a << " , " << b << ") = " << g << '\n';
+	// now we know n2 > n1
+	rm = n2%n1;
+	while(rm != 0)
+	{
+		n2 = n1;
+		n1 = rm;
+		rm = n2%n1;
+	};
 	
-	return 0;
+	
+	return n1;
+	
 }
